-VPIP and PFR print cosmetics -3BET stat added
