+++ CHANGELOG.MD
@@ -1,0 +1,3 @@
+1.0.1
+    - STATS
+        - VPIP, PFR cosmetics; 3BET, AF stats added
